fix(trie): address RocksDB review feedback
